Refactoring

+++ flower-shop-backend/src/main/java/com/pro100user/decorshopbackend/service/impl/ProductServiceImpl.java
@@ -64,6 +64,10 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public boolean delete(Long productId) {
+        Product product = productRepository.findById(productId).orElseThrow();
+        if(!product.getBaskets().isEmpty() || !product.getOrders().isEmpty()) {
+            throw new IllegalArgumentException("Видалення не доступне, так як цей товар використовується!");
+        }
         productRepository.deleteById(productId);
         return true;
     }
